Write realization, many fixes

+++ include/pt/object/sky.h
@@ -13,8 +13,8 @@ namespace pt
 
 		bool intersect(const Ray& ray, 
 					   Intersection& inter, 
-					   float tMin, 
-					   float tMax) const;
+					   double tMin, 
+					   double tMax) const;
 		ScatterType scatter(const Ray& ray,
 							const Intersection& inter,
 							Color& clrAbsorbtion,
